+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+## Version 1.0.0-RC2
+_2016-03-30_
+- **[IMPROVEMENTS]** Update assets for Wordpress publication
+- **[IMPROVEMENTS]** Add publish.sh file for publishing version on the SVN repo.
+
 ## Version 1.0.0-RC1
 _2016-03-24_
 - **[IMPROVEMENTS]** : Add Readme.txt for wordpress repo and the icon image.